Homogeneization of autoencoder training
