+++ clinicadl/clinicadl/tools/deep_learning/iotools.py
@@ -267,59 +267,6 @@ def read_json(options, task_type, json_path=None, test=False):
     return options
 
 
-def visualize_subject(decoder, dataloader, visualization_path, options, epoch=None, save_input=False, subject_index=0):
-    from os import path, makedirs, pardir
-    import nibabel as nib
-    import numpy as np
-    import torch
-    from .data import MinMaxNormalization
-
-    if not path.exists(visualization_path):
-        makedirs(visualization_path)
-
-    dataset = dataloader.dataset
-    data = dataset[subject_index]
-    image_path = data['image_path']
-
-    # TODO: Change nifti path
-    nii_path, _ = path.splitext(image_path)
-    nii_path += '.nii.gz'
-
-    if not path.exists(nii_path):
-        nii_path = path.join(
-            path.dirname(image_path),
-            pardir, pardir, pardir,
-            't1_linear',
-            path.basename(image_path)
-        )
-        nii_path, _ = path.splitext(nii_path)
-        nii_path += '.nii.gz'
-    
-    input_nii = nib.load(nii_path)
-    input_np = input_nii.get_data().astype(float)
-    np.nan_to_num(input_np, copy=False)
-    input_pt = torch.from_numpy(input_np).unsqueeze(0).unsqueeze(0).float()
-    if options.minmaxnormalization:
-        transform = MinMaxNormalization()
-        input_pt = transform(input_pt)
-
-    if options.gpu:
-        input_pt = input_pt.cuda()
-
-    output_pt = decoder(input_pt)
-
-    output_np = output_pt.detach().cpu().numpy()[0][0]
-    output_nii = nib.Nifti1Image(output_np, affine=input_nii.affine)
-
-    if save_input:
-        nib.save(input_nii, path.join(visualization_path, 'input.nii'))
-
-    if epoch is None:
-        nib.save(output_nii, path.join(visualization_path, 'output.nii'))
-    else:
-        nib.save(output_nii, path.join(visualization_path, 'epoch-' + str(epoch) + '.nii'))
-
-
 def memReport():
     import gc
     import torch
